State System
Assets that can be created include state components, states, decisions, and transitions. Transitions are made up of states and a decision, states are made up of any number of state components. Transitions can be used on a brain component attached to a game object to dynamically build a state machine.